fix : put the field mask filter in post filters instead of filters
