+++ crates/crabe_filter/src/post_filter.rs
@@ -2,6 +2,7 @@ pub mod ball;
 pub mod geometry;
 pub mod robot;
 pub mod game_controller;
+pub mod field_mask;
 
 use crate::data::FilterData;
 use crabe_framework::data::world::World;
