+++ crates/crabe_io/src/bin/vision_log.rs
@@ -3,7 +3,7 @@ use crabe_protocol::protobuf::vision_packet::SslWrapperPacket;
 use std::net::Ipv4Addr;
 
 fn main() {
-    let mut vision = MulticastUDPReceiver::new(Ipv4Addr::new(224, 5, 23, 2), 10020)
+    let mut vision = MulticastUDPReceiver::new(Ipv4Addr::new(224, 5, 23, 2), 10006)
         .expect("Error to create Vision UDP Receiver");
     loop {
         if let Some(packet) = vision.receive::<SslWrapperPacket>() {
